Share some utilities between Gateways
